gh-801: clean up restapi package / split up configure_weviate.go
It was becoming increasingly hard to work with this file, so this splits
it up into smaller files. This does not solve all the problems with it
yet, but at least it makes it a bit nicer to work with.
