+++ restapi/handlers_misc.go
@@ -1,0 +1,58 @@
+package restapi
+
+import (
+	"github.com/creativesoftwarefdn/weaviate/database/schema"
+	"github.com/creativesoftwarefdn/weaviate/models"
+	"github.com/creativesoftwarefdn/weaviate/network/common/peers"
+	"github.com/creativesoftwarefdn/weaviate/restapi/operations"
+	"github.com/creativesoftwarefdn/weaviate/restapi/operations/meta"
+	"github.com/creativesoftwarefdn/weaviate/restapi/operations/p2_p"
+	middleware "github.com/go-openapi/runtime/middleware"
+)
+
+func setupMiscHandlers(api *operations.WeaviateAPI) {
+	api.MetaWeaviateMetaGetHandler = meta.WeaviateMetaGetHandlerFunc(func(params meta.WeaviateMetaGetParams, principal *models.Principal) middleware.Responder {
+		dbLock, err := db.ConnectorLock()
+		if err != nil {
+			return meta.NewWeaviateMetaGetInternalServerError().WithPayload(errPayloadFromSingleErr(err))
+		}
+		defer dbLock.Unlock()
+		databaseSchema := schema.HackFromDatabaseSchema(dbLock.GetSchema())
+		// Create response object
+		metaResponse := &models.Meta{}
+
+		// Set the response object's values
+		metaResponse.Hostname = serverConfig.GetHostAddress()
+		metaResponse.ActionsSchema = databaseSchema.ActionSchema.Schema
+		metaResponse.ThingsSchema = databaseSchema.ThingSchema.Schema
+
+		return meta.NewWeaviateMetaGetOK().WithPayload(metaResponse)
+	})
+
+	api.P2PWeaviateP2pGenesisUpdateHandler = p2_p.WeaviateP2pGenesisUpdateHandlerFunc(func(params p2_p.WeaviateP2pGenesisUpdateParams) middleware.Responder {
+		newPeers := make([]peers.Peer, 0)
+
+		for _, genesisPeer := range params.Peers {
+			peer := peers.Peer{
+				ID:         genesisPeer.ID,
+				Name:       genesisPeer.Name,
+				URI:        genesisPeer.URI,
+				SchemaHash: genesisPeer.SchemaHash,
+			}
+
+			newPeers = append(newPeers, peer)
+		}
+
+		err := network.UpdatePeers(newPeers)
+
+		if err == nil {
+			return p2_p.NewWeaviateP2pGenesisUpdateOK()
+		}
+		return p2_p.NewWeaviateP2pGenesisUpdateInternalServerError()
+	})
+
+	api.P2PWeaviateP2pHealthHandler = p2_p.WeaviateP2pHealthHandlerFunc(func(params p2_p.WeaviateP2pHealthParams) middleware.Responder {
+		// For now, always just return success.
+		return middleware.NotImplemented("operation P2PWeaviateP2pHealth has not yet been implemented")
+	})
+}
