+++ restapi/helpers.go
@@ -1,0 +1,70 @@
+package restapi
+
+import (
+	"fmt"
+	"math"
+
+	"github.com/creativesoftwarefdn/weaviate/models"
+	"github.com/go-openapi/strfmt"
+)
+
+const pageOverride int = 1
+
+// getLimit returns the maximized limit
+func getLimit(paramMaxResults *int64) int {
+	maxResults := serverConfig.Environment.Limit
+	// Get the max results from params, if exists
+	if paramMaxResults != nil {
+		maxResults = *paramMaxResults
+	}
+
+	// Max results form URL, otherwise max = config.Limit.
+	return int(math.Min(float64(maxResults), float64(serverConfig.Environment.Limit)))
+}
+
+// getPage returns the page if set
+func getPage(paramPage *int64) int {
+	page := int64(pageOverride)
+	// Get the page from params, if exists
+	if paramPage != nil {
+		page = *paramPage
+	}
+
+	// Page form URL, otherwise max = config.Limit.
+	return int(page)
+}
+
+func generateMultipleRefObject(keyIDs []strfmt.UUID) models.MultipleRef {
+	// Init the response
+	refs := models.MultipleRef{}
+
+	// Generate SingleRefs
+	for _, keyID := range keyIDs {
+		refs = append(refs, &models.SingleRef{
+			NrDollarCref: strfmt.URI(keyID),
+		})
+	}
+
+	return refs
+}
+
+// createErrorResponseObject is a common function to create an error response
+func createErrorResponseObject(messages ...string) *models.ErrorResponse {
+	// Initialize return value
+	er := &models.ErrorResponse{}
+
+	// appends all error messages to the error
+	for _, message := range messages {
+		er.Error = append(er.Error, &models.ErrorResponseErrorItems0{
+			Message: message,
+		})
+	}
+
+	return er
+}
+
+func errPayloadFromSingleErr(err error) *models.ErrorResponse {
+	return &models.ErrorResponse{Error: []*models.ErrorResponseErrorItems0{{
+		Message: fmt.Sprintf("%s", err),
+	}}}
+}
